# openapi.yaml
...
openapi: 3.0.3
info:
  title: Oficina MVP API
  description: Tech Challenge 15SOAT — Sistema de Gestão de Oficina Mecânica (fases
    1 e 2).
  version: 1.0.0
...
paths:
...
  /clientes:
    get:
      tags:
      - Clientes
      summary: Lista todos os clientes
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClienteResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
    post:
      tags:
      - Clientes
      summary: Cadastra um novo cliente
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CadastrarClienteRequest"
      responses:
        "201":
          description: Cliente criado
        "400":
          description: Documento ou payload inválido
        "409":
          description: Documento já cadastrado
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
  /clientes/{id}:
    get:
      tags:
      - Clientes
      summary: Busca um cliente pelo ID
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "404":
          description: Cliente não encontrado
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
    put:
      tags:
      - Clientes
      summary: Atualiza nome/contato de um cliente
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AtualizarClienteRequest"
      responses:
        "404":
          description: Cliente não encontrado
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
    delete:
      tags:
      - Clientes
      summary: Remove um cliente
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "204":
          description: Removido
        "404":
          description: Cliente não encontrado
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
...
  /veiculos/cliente/{clienteId}:
    get:
      tags:
      - Veículos
      summary: Lista veículos de um cliente
      parameters:
      - name: clienteId
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VeiculoResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
...
components:
  schemas:
...
    AtualizarClienteRequest:
      required:
      - nome
      type: object
      properties:
        nome:
          description: Nome completo do cliente
          maxLength: 150
          pattern: \S
          type: string
          example: João da Silva
        email:
          description: Email do cliente
          maxLength: 150
          type: string
          example: joao.silva@example.com
        telefone:
          description: Telefone do cliente
          maxLength: 30
          type: string
          example: "11999999999"
...
    CadastrarClienteRequest:
      required:
      - nome
      - documento
      type: object
      properties:
        nome:
          description: Nome completo do cliente
          maxLength: 150
          pattern: \S
          type: string
          example: João da Silva
        documento:
          description: CPF ou CNPJ do cliente
          maxLength: 18
          pattern: \S
          type: string
          example: "12345678900"
        email:
          description: Email do cliente
          maxLength: 150
          type: string
          example: joao.silva@example.com
        telefone:
          description: Telefone do cliente
          maxLength: 30
          type: string
          example: "11999999999"
...
    ClienteResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        nome:
          type: string
        documento:
          type: string
        tipoDocumento:
          type: string
        email:
          type: string
        telefone:
          type: string
...
    UUID:
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
      type: string
...
    VeiculoResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        placa:
          type: string
        marca:
          type: string
        modelo:
          type: string
        ano:
          format: int32
          type: integer
        clienteId:
          $ref: "#/components/schemas/UUID"
  securitySchemes:
    BearerAuth:
      type: http
      description: "JWT emitido pelo POST /auth/login (cole apenas o token, sem prefixo)."
      scheme: bearer
      bearerFormat: JWT
    SecurityScheme:
      type: http
      description: Authentication
      scheme: bearer
      bearerFormat: JWT